# docs/openapi/job-autopilot.yaml
openapi: 3.0.3
info:
  title: Job Autopilot Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://staging.api.jobautopilot.ai
    description: Staging
  - url: https://api.jobautopilot.ai
    description: Production
paths:
  /health:
    get:
      summary: Platform health check
      responses:
        '200':
          description: OK
  /v1/jobs/intake:
    post:
      summary: Intake job link or ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobIntakeRequest'
      responses:
        '200':
          description: Job parsed
  /v1/workflows/tailor:
    post:
      summary: Tailor resume and cover letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TailorRequest'
      responses:
        '200':
          description: Workflow queued
  /v1/jobs/fetch:
    post:
      summary: Fetch job description from link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobFetchRequest'
      responses:
        '200':
          description: Job description fetched
  /v1/channels/intake:
    post:
      summary: Intake Telegram/WhatsApp message for OpenClaw workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelIntakeRequest'
      responses:
        '200':
          description: OpenClaw workflow started
components:
  schemas:
    JobIntakeRequest:
      type: object
      properties:
        source:
          type: string
          example: job_link
        job_link:
          type: string
          format: uri
        job_id:
          type: string
        user_id:
          type: string
      required:
        - source
    TailorRequest:
      type: object
      properties:
        job_id:
          type: string
        resume_id:
          type: string
        user_id:
          type: string
        include_cover_letter:
          type: boolean
      required:
        - job_id
        - resume_id
        - user_id
    JobFetchRequest:
      type: object
      properties:
        job_link:
          type: string
          format: uri
        user_id:
          type: string
      required:
        - job_link
    ChannelIntakeRequest:
      type: object
      properties:
        user_id:
          type: string
        user_handle:
          type: string
        channel:
          type: string
          enum: [telegram, whatsapp]
        message:
          type: string
        job_link:
          type: string
          format: uri
        job_description:
          type: string
        name:
          type: string
        job_title:
          type: string
        company:
          type: string
        template_name:
          type: string
      required:
        - user_id
        - user_handle
        - channel
        - message